refactor: Cars, car 객체에 행위를 위임하도록 변경
