+++ src/main/java/step2/domain/Car.java
@@ -1,6 +1,8 @@
 package step2.domain;
 
 public class Car {
+    public static final int MIN_RANDOM_NUMBER = 4;
+
     private int position;
     private String name;
 
@@ -17,7 +19,23 @@ public class Car {
         return name;
     }
 
-    public void setPosition(int position) {
-        this.position = position;
+    public void move(int random) {
+        this.position = tryMove(random);
+    }
+
+    private int tryMove(int randomNumber) {
+        int movedPosition = this.position;
+
+        movedPosition = moveForward(randomNumber, movedPosition);
+
+        return movedPosition;
+    }
+
+    private int moveForward(int number, int position) {
+        if (number >= MIN_RANDOM_NUMBER) {
+            position = position + 1;
+        }
+
+        return position;
     }
 }
