+++ src/main/java/step2/domain/RaceController.java
@@ -16,7 +16,7 @@ public class RaceController {
         ResultView.printStartMessage(racingGame.getCars().getCarNames());
 
         for (int i = 0; i < tryCount; i++) {
-            ResultView.print(racingGame.move());
+            ResultView.print(racingGame.race());
         }
 
         List<String> winners = racingGame.retrieveWinners();
