+++ src/main/java/step2/domain/RacingGame.java
@@ -1,10 +1,7 @@
 package step2.domain;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
-
-import static java.util.stream.Collectors.toList;
 
 public class RacingGame {
     private static final String CAR_NAME_DELIMITER = ",";
@@ -26,64 +23,22 @@ public class RacingGame {
     }
 
     public List<String> retrieveWinners() {
-        List<String> winners = new ArrayList<>();
-        List<Car> carCollection = this.cars.getCars();
-
-        int maxPosition = getMaxPosition(carCollection);
-
-        return carCollection.stream()
-                .map(car -> getWinner(car, maxPosition))
-                .filter(winner -> !winner.equals(""))
-                .collect(toList());
+        return this.cars.retrieveWinners();
     }
 
-    public int getMaxPosition(List<Car> carCollection) {
-        int maxPosition = 0;
-
-        for (Car car : carCollection) {
-            maxPosition = findMaxPosition(car, maxPosition);
-        }
-
-        return maxPosition;
-    }
-
-    private int findMaxPosition(Car car, int maxPosition) {
-        return (car.getPosition() > maxPosition) ? car.getPosition() : maxPosition;
-    }
-
-    public String getWinner(Car car, int maxPosition) {
-        String winnerName = "";
-
-        if (car.getPosition() == maxPosition) {
-            winnerName = car.getName();
-        }
-        return winnerName;
-    }
-
-
-    public Cars move() {
-        int length = numberOfCars;
-        int position;
-        int resultPosition;
+    public Cars race() {
+        int length = this.numberOfCars;
+        int random;
 
         List<Car> carCollection = cars.getCars();
 
         for (int i = 0; i < length; i++) {
-            position = carCollection.get(i).getPosition();
-            resultPosition = iterateCarMove(position);
-            carCollection.get(i).setPosition(resultPosition);
+            random = generateRandom();
+            Car car = carCollection.get(i);
+            car.move(random);
         }
 
         return this.cars;
-    }
-
-    public int iterateCarMove(int position) {
-        int randomNumber = generateRandom();
-        int movedPosition = position;
-
-        movedPosition = moveForward(randomNumber, movedPosition);
-
-        return movedPosition;
     }
 
     public int generateRandom() {
@@ -91,14 +46,6 @@ public class RacingGame {
         int randomNumber = random.nextInt(10);
 
         return randomNumber;
-    }
-
-    public int moveForward(int number, int position) {
-        if (number >= 4) {
-            position = position + 1;
-        }
-
-        return position;
     }
 
     public Cars getCars() {
